implement curve + msm macros
