+++ wrappers/rust/icicle-curves/icicle-bn254/src/lib.rs
@@ -1,3 +1,6 @@
+#[macro_use(impl_curve)]
+extern crate icicle_core;
+
 pub mod curve;
 pub mod msm;
 pub mod ntt;
